bad: give up on tinyfpga_usb_serial, for now
Turns out even in the makefile version, it isn't closing timing at 48MHz
on my build. Somewhere, it's getting a default of 12MHz (I guess because
pll.v isn't present?), and only getting up to 31MHz or so.

I'm seeing the same results when building here, but because we've
"correctly" specified the clock in the .pcf, nextpnr complains that it
can't close timing.