fix: debug FIFO issues in http_server
According to the waves, the FIFO inputs are alright in this
configuration. Something is getting wonky when reading the outputs (I
had feared the r_en cycle comment may be the problem.)

I pulled everything down to the "sync" domain, which may be helping?
This at least passes the tests.

Oh, wow, and it passes timing! So maybe this is fine?

Yeah, this works on the device, with one (two) bug(s):
it always reads "0 seconds since startup", and it has some not-great
buffering. Oh well, it's progress!