Add TLP rules
I had forgotten this setup step, and had very poor notes on it.

Laptop power savings can suspend USB devices at inconvenient times. The
USB softcore we're working with on Fomu doesn't play nicely with
suspend/resume.

Add a dropin to tlp(8) to disable autosuspend for the Fomu- both
bootloader (which actually works fine with it) and serial device (which
doesn't).